+++ MQ2ItemDisplay/ISXEQItemDisplay.cpp
@@ -37,9 +37,9 @@ HISXSERVICE hEQSpawnService=0;
 HISXSERVICE hEQZoneService=0;
 
 // Forward declarations of callbacks
-void __cdecl PulseService(bool Broadcast, unsigned long MSG, void *lpData);
-void __cdecl MemoryService(bool Broadcast, unsigned long MSG, void *lpData);
-void __cdecl ServicesService(bool Broadcast, unsigned long MSG, void *lpData);
+void __cdecl PulseService(bool Broadcast, unsigned int MSG, void *lpData);
+void __cdecl MemoryService(bool Broadcast, unsigned int MSG, void *lpData);
+void __cdecl ServicesService(bool Broadcast, unsigned int MSG, void *lpData);
 
 
 // Initialize is called by Inner Space when the extension should initialize.
@@ -184,7 +184,7 @@ void ISXEQItemDisplay::UnRegisterServices()
 //		pISInterface->ShutdownService(this,hISXEQItemDisplayService);
 }
 
-void __cdecl MemoryService(bool Broadcast, unsigned long MSG, void *lpData)
+void __cdecl MemoryService(bool Broadcast, unsigned int MSG, void *lpData)
 {
 	// no messages are currently associated with this service (other than
 	// system messages such as client disconnect), so do nothing.
